+++ src/Containers/Interfaces/IStackInventory.cs
@@ -68,7 +68,7 @@ namespace TheChest.Inventories.Containers.Interfaces
         /// </summary>
         /// <param name="item">The item to evaluate to add to the inventory.</param>
         /// <returns>true if the <paramref name="item"/> can be added; otherwise, false.</returns>
-        [Obsolete("Use CanAddAt(params T[]) instead to check if the item can be added.")]
+        [Obsolete("Use CanAddAt(params T[]) instead. This method will be removed in future versions.")]
         bool CanAdd(T item);
         /// <summary>
         /// Checks if <paramref name="items"/> can be added to any slot on inventory.
@@ -88,6 +88,7 @@ namespace TheChest.Inventories.Containers.Interfaces
         /// </summary>
         /// <param name="item">item to be added</param>
         /// <returns><see langword="true"/> if is possible to add <paramref name="item"/></returns>
+        [Obsolete("Use Add(params T[]) instead. This method will be removed in future versions.")]
         bool Add(T item);
 
         /// <summary>